Update index.html
Prompt : there's a bug, the `showReel` function should also be able to accept `url` as a parameter, and if the parameter isn't available, then get the url from the input box

Response : 
I apologize for the confusion. I misunderstood your question earlier. Here's the updated `showReel` function that accepts a `url` parameter:

Now, you can call the `showReel` function with a url parameter like this:

```javascript
showReel('https://www.instagram.com/reel/12345/');
```

If the `url` parameter is not provided, the function will use the value from the input box.