Update index.html
Prompt 1 : add a "Copy URL" button that copies the current browser URL to the clipboard. This code should work for both desktop and mobile devices.

Response : 
Add the "Copy URL" button to the HTML:
Now, when you click the "Copy URL" button, the current browser URL will be copied to the clipboard.